+++ company-capf.el
@@ -102,6 +102,11 @@ Requires Emacs 24.1 or newer."
     (`location
      (let ((f (plist-get (nthcdr 4 (company--capf-data)) :company-location)))
        (when f (funcall f arg))))
+    (`annotation
+     (save-excursion
+       (goto-char company-point)
+       (let ((f (plist-get (nthcdr 4 (company--capf-data)) :annotation-function)))
+         (when f (funcall f arg)))))
     (`require-match
      (plist-get (nthcdr 4 (company--capf-data)) :company-require-match))
     (`init nil)      ;Don't bother: plenty of other ways to initialize the code.